Updated "profiles" path construction
Changed from "user.dir" to the Processing function sketchPath(""). The
former will be the directory the program is running from whereas the
latter is the actual location of the file that is being run, which is
what we want because the "profiles" directory is relative to the
osPID_FrontEnd.pde file.